Refactored the naming convention more acceptable
